requirements.txt
# PDF processing libraries
pdfplumber==0.10.4
PyPDF2==3.0.1

# OCR and image processing
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.17.0

# Optional advanced image processing
opencv-python==4.9.0.80

# Type checking and development
mypy==1.8.0

# Logging 